Create single command and split fetchCommit to separate module
