+++ packages/rn-changelog-generator/src/generator.ts
@@ -7,10 +7,8 @@ import fs from "fs";
 import chalk from "chalk";
 import pLimit from "p-limit";
 import deepmerge from "deepmerge";
-import https from "https";
 import child_process from "child_process";
-import { IncomingHttpHeaders } from "http";
-import { GenerateArgs } from "./utils/args";
+import fetchCommits, { Commit } from "./utils/fetchCommits";
 
 const execFile = util.promisify(child_process.execFile);
 
@@ -45,104 +43,9 @@ const CHANGELOG_LINE_REGEXP = new RegExp(
   "i"
 );
 
-export interface Commit {
-  sha: string;
-  commit: { message: string };
-  author?: { login: string };
-}
-
 export type PlatformChanges = Record<typeof CHANGE_CATEGORY[number], string[]>;
 
 export type Changes = Record<typeof CHANGE_TYPE[number], PlatformChanges>;
-
-//#region NETWORK
-//*****************************************************************************
-
-function fetchJSON<T>(token: string | null, path: string) {
-  const host = "api.github.com";
-  console.warn(chalk.yellow(`https://${host}${path}`));
-  return new Promise<{ json: T; headers: IncomingHttpHeaders }>(
-    (resolve, reject) => {
-      let data = "";
-
-      https
-        .get({
-          host,
-          path,
-          headers: {
-            ...(token != null ? { Authorization: `token ${token}` } : null),
-            "User-Agent":
-              "https://github.com/react-native-community/releases/blob/master/scripts/changelog-generator.js",
-          },
-        })
-        .on("response", (response) => {
-          if (response.statusCode !== 200) {
-            return reject(
-              new Error(`[!] Got HTTP status: ${response.statusCode}`)
-            );
-          }
-
-          response.on("data", (chunk) => {
-            data += chunk;
-          });
-
-          response.on("end", () => {
-            try {
-              resolve({ json: JSON.parse(data), headers: response.headers });
-            } catch (e) {
-              reject(e);
-            }
-          });
-
-          response.on("error", (error) => {
-            reject(error);
-          });
-        });
-    }
-  );
-}
-
-export function fetchCommits(
-  token: string | null,
-  base: string,
-  compare: string
-) {
-  console.warn(chalk.green("Fetch commit data"));
-  console.group();
-  const commits: Commit[] = [];
-  let page = 1;
-  return new Promise<Commit[]>((resolve, reject) => {
-    const fetchPage = () => {
-      fetchJSON<Commit[]>(
-        token,
-        `/repos/facebook/react-native/commits?sha=${compare}&page=${page++}`
-      )
-        .then(({ json, headers }) => {
-          for (const commit of json) {
-            commits.push(commit);
-            if (commit.sha === base) {
-              console.groupEnd();
-              return resolve(commits);
-            }
-          }
-          if (!(headers["link"] as string).includes("next")) {
-            throw new Error(
-              "Did not find commit after paging through all commits"
-            );
-          }
-          setImmediate(fetchPage);
-        })
-        .catch((e) => {
-          console.groupEnd();
-          reject(e);
-        });
-    };
-    fetchPage();
-  });
-}
-
-//*****************************************************************************
-//#endregion
 
 //#region FILTER COMMITS
 //*****************************************************************************
@@ -760,7 +663,17 @@ export async function run(
   return buildMarkDown(options.compare, changes);
 }
 
-export default function (argv: GenerateArgs) {
+interface GenerateArgs {
+  base: string;
+  compare: string;
+  repo: string;
+  changelog: string;
+  token: string | null;
+  maxWorkers: number;
+  verbose: boolean;
+}
+
+function handler(argv: GenerateArgs) {
   const gitDir = path.join(argv.repo, ".git");
   git(gitDir, "rev-parse")
     .catch(() => {
@@ -779,6 +692,62 @@ export default function (argv: GenerateArgs) {
       process.exit(1);
     });
 }
+
+export default {
+  handler,
+  args: {
+    base: {
+      alias: "b",
+      string: true,
+      describe:
+        "The base branch/tag/commit to compare against (most likely the previous stable version)",
+      demandOption: true,
+    },
+    compare: {
+      alias: "c",
+      string: true,
+      describe:
+        "The new version branch/tag/commit (most likely the latest release candidate)",
+      demandOption: true,
+    },
+    repo: {
+      alias: "r",
+      string: true,
+      describe: "The path to an up-to-date clone of the react-native repo",
+      demandOption: true,
+    },
+    changelog: {
+      alias: "f",
+      string: true,
+      describe: "The path to the existing CHANGELOG.md file",
+      demandOption: true,
+      default: path.resolve(__dirname, "../CHANGELOG.md"),
+    },
+    token: {
+      alias: "t",
+      string: true,
+      describe:
+        "A GitHub token that has `public_repo` access (generate at https://github.com/settings/tokens)",
+      demandOption: false,
+      default: null,
+    },
+    maxWorkers: {
+      alias: "w",
+      number: true,
+      describe:
+        "Specifies the maximum number of concurrent sub-processes that will be spawned",
+      default: 10,
+    },
+    verbose: {
+      alias: "v",
+      boolean: true,
+      describe:
+        "Verbose listing, includes internal changes as well as public-facing changes",
+      demandOption: false,
+      default: false,
+    },
+  },
+};
 
 //*****************************************************************************
 //#endregion
