--- packages/rn-changelog-generator/src/utils/args.ts
@@ -1,9 +1,0 @@
-export interface GenerateArgs {
-  base: string;
-  compare: string;
-  repo: string;
-  changelog: string;
-  token: string | null;
-  maxWorkers: number;
-  verbose: boolean;
-}
